rename macro
