+++ .travis.yml
@@ -5,5 +5,6 @@ python:
   - "3.6"
   - "3.7"
   - "3.8"
+  - "3.9"
 script:
   - python tests.py
